GUAC-1132: Split non-database permission service functions into interfaces.
