+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/permission/ConnectionPermissionService.java
@@ -35,7 +35,7 @@ import org.glyptodon.guacamole.auth.jdbc.user.ModeledUser;
  *
  * @author Michael Jumper
  */
-public class ConnectionPermissionService extends ObjectPermissionService {
+public class ConnectionPermissionService extends ModeledObjectPermissionService {
 
     /**
      * Mapper for connection permissions.
