+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/permission/UserPermissionService.java
@@ -35,7 +35,7 @@ import org.glyptodon.guacamole.auth.jdbc.user.ModeledUser;
  *
  * @author Michael Jumper
  */
-public class UserPermissionService extends ObjectPermissionService {
+public class UserPermissionService extends ModeledObjectPermissionService {
 
     /**
      * Mapper for user permissions.
